+++ Assets/Scripts/InGame/UI.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: bf047a6bbe50845bd92a169198ae8859
+guid: c533fde24e74dd740a84c3f39bc8f89a
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
