+++ Assets/Scripts/InGame/UI/BarPlotComponent.cs
