+++ Assets/Scripts/InGame/UI/PalatteInk.cs.meta
