+++ Assets/Scripts/Ink.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: ab9898a5e2a10f54f87cb2ee98598a3a
+guid: 721c93825ee3afb48825bcf6f2ea78b4
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
