+++ Assets/Scripts/Ink/AreaCalculation.cs
