+++ Assets/Scripts/Ink/ColorType.cs
