+++ Assets/Scripts/Stage.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 87d30d12c4da1ba41b1c86007031eeee
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
